backend/assembly-images-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=1.0.0
pydantic>=2.0.0
pillow>=10.0.0
pillow-avif-plugin>=1.4.0
python-dotenv>=1.0.0
